Clean up and rename GuiEditMenuTier
